move runtimehook into mydag
